Menubar, Border- & FlowPane
Wrote a menu bar that rests at top position of a border pane, which
holds a flowpane in its center position. The flowpane will contain
children nodes for dark/light colored squares and different game pieces,
and wrap the nodes in each row when a set horizontal spacing is reached.