Board grid GUI
Began implementation of the graphical (user interactive) chess board,
which is represented by a JavaFX GridPane that holds 8*8 StackPanes
representing each square which in turn will "stack" a Rectangle and a
Canvas object, or maybe just a Canvas object that draws the correct
background with a chess game piece overlapping. If this is the case, no
StackPane will be needed for the Square as it wil suffice with the
Canvas. Time will tell.