Highlight legal moves
Implemented methods for highlighting a selected piece's legal moves on
the chess board by coloring the "background" square of the square stack
in a different color. The algorithm runs through every legal move for
the selected piece and checks which possible move destinations
corresponds to each board square. When a successful match is found, the
square gets a different color.